test: separate tests by function
