+++ test/globalSetup.ts
@@ -4,7 +4,12 @@ import { cwd } from 'node:process';
 import { readFile } from 'node:fs/promises';
 
 // constants
-import { GENESIS_ACCOUNT_ID, SOCIAL_CONTRACT_ACCOUNT_ID } from './constants';
+import {
+  CAPTAIN_KIRK_ACCOUNT_ID,
+  GENESIS_ACCOUNT_ID,
+  SOCIAL_CONTRACT_ACCOUNT_ID,
+  SPOCK_ACCOUNT_ID,
+} from './constants';
 
 // helpers
 import convertNEARToYoctoNEAR from './helpers/convertNEARToYoctoNEAR';
@@ -13,29 +18,50 @@ import createTestAccount from './helpers/createTestAccount';
 
 export default async function globalSetup() {
   const _functionName = 'globalSetup';
+  const connection = await createNearConnection({
+    credentialsDir: resolve(cwd(), 'test', 'credentials'),
+  });
   const contract = await readFile(
     resolve(cwd(), 'test', 'contracts', 'social_db.wasm')
   );
-  const near = await createNearConnection({
-    credentialsDir: resolve(cwd(), 'test', 'credentials'),
-  });
+  const networkId = 'localnet';
   let contractAccountPublicKey: utils.PublicKey;
   let contractAccount: Account;
-  let creatorAccount: Account;
+  let genesisAccount: Account;
 
-  creatorAccount = await near.account(GENESIS_ACCOUNT_ID);
-  contractAccountPublicKey = await near.connection.signer.getPublicKey(
+  genesisAccount = await connection.account(GENESIS_ACCOUNT_ID);
+  contractAccountPublicKey = await connection.connection.signer.getPublicKey(
     SOCIAL_CONTRACT_ACCOUNT_ID,
-    'localnet'
+    networkId
   );
+
+  // create known accounts
+  await createTestAccount({
+    connection,
+    creatorAccount: genesisAccount,
+    newAccountID: CAPTAIN_KIRK_ACCOUNT_ID,
+    newAccountPublicKey: await connection.connection.signer.getPublicKey(
+      CAPTAIN_KIRK_ACCOUNT_ID,
+      networkId
+    ),
+  });
+  await createTestAccount({
+    connection,
+    creatorAccount: genesisAccount,
+    newAccountID: SPOCK_ACCOUNT_ID,
+    newAccountPublicKey: await connection.connection.signer.getPublicKey(
+      SPOCK_ACCOUNT_ID,
+      networkId
+    ),
+  });
 
   // create the contract account
   contractAccount = await createTestAccount({
-    creatorAccount,
+    creatorAccount: genesisAccount,
     initialBalanceInAtomicUnits: convertNEARToYoctoNEAR(BigInt('10')),
     newAccountID: SOCIAL_CONTRACT_ACCOUNT_ID,
     newAccountPublicKey: contractAccountPublicKey,
-    nearConnection: near,
+    connection: connection,
   });
 
   // deploy the account
@@ -43,11 +69,11 @@ export default async function globalSetup() {
 
   try {
     // initialize the contract
-    await creatorAccount.functionCall({
+    await genesisAccount.functionCall({
       contractId: contractAccount.accountId,
       methodName: 'new',
     });
   } catch (error) {
-    console.error(`${_functionName}:`, error);
+    console.debug(`${_functionName}:`, error);
   }
 }
